+++ config-server/src/main/java/cz/upce/votingsystemapplication/VotingSystemApplicationConfigServer.java
@@ -1,0 +1,15 @@
+package cz.upce.votingsystemapplication;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.config.server.EnableConfigServer;
+
+@EnableConfigServer
+@SpringBootApplication
+public class VotingSystemApplicationConfigServer {
+
+    public static void main(String[] args) {
+        SpringApplication.run(VotingSystemApplicationConfigServer.class, args);
+    }
+
+}
