+++ voting-system-application/src/main/java/cz/upce/votingsystemapplication/VotingSystemApplication.java
@@ -1,0 +1,15 @@
+package cz.upce.votingsystemapplication;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@EnableDiscoveryClient
+@SpringBootApplication
+public class VotingSystemApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(VotingSystemApplication.class, args);
+    }
+
+}
