+++ voting-system-application/src/main/java/cz/upce/votingsystemapplication/dao/TagDAO.java
@@ -1,0 +1,12 @@
+package cz.upce.votingsystemapplication.dao;
+
+import cz.upce.votingsystemapplication.model.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TagDAO  extends JpaRepository<Tag, Long> {
+//     Tag getTagOnSuggestion(Suggestion suggestionId);
+     Tag findOneBySuggestion_Id(Long suggestionId);
+    }
+
