+++ voting-system-application/src/main/java/cz/upce/votingsystemapplication/dao/UserDao.java
@@ -1,0 +1,8 @@
+package cz.upce.votingsystemapplication.dao;
+
+import cz.upce.votingsystemapplication.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserDao extends JpaRepository<User,Long> {
+
+}
