+++ voting-system-application/src/main/java/cz/upce/votingsystemapplication/dto/MeetingDto.java
@@ -1,0 +1,23 @@
+package cz.upce.votingsystemapplication.dto;
+
+import java.sql.Timestamp;
+import java.util.List;
+
+public class MeetingDto extends MeetingForSuggestionDto {
+
+    private List<SuggestionForMeetingDto> suggestions;
+
+    public MeetingDto(long id, Timestamp start, List<SuggestionForMeetingDto> suggestions) {
+        super(id, start);
+
+        this.suggestions = suggestions;
+    }
+
+    public List<SuggestionForMeetingDto> getSuggestions() {
+        return suggestions;
+    }
+
+    public void setSuggestions(List<SuggestionForMeetingDto> suggestions) {
+        this.suggestions = suggestions;
+    }
+}
