+++ voting-system-application/src/main/java/cz/upce/votingsystemapplication/service/TagService.java
@@ -1,0 +1,42 @@
+package cz.upce.votingsystemapplication.service;
+
+import cz.upce.votingsystemapplication.dao.TagDAO;
+import cz.upce.votingsystemapplication.dto.TagDto;
+import cz.upce.votingsystemapplication.model.Tag;
+import org.modelmapper.ModelMapper;
+import org.modelmapper.TypeToken;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class TagService {
+
+    @Autowired
+    private TagDAO tagDAO;
+
+    ModelMapper modelMapper=new ModelMapper();
+
+    public void addTag(Tag tag) {
+        tagDAO.save(tag);
+    }
+
+    public TagDto getTag(Long id) {
+        Tag tag = tagDAO.getOne(id);
+        return modelMapper.map(tag, TagDto.class);
+    }
+
+    public List<TagDto> getAllTag() {
+        List<Tag> all = tagDAO.findAll();
+        return modelMapper.map(all,new TypeToken<List<TagDto>>() {}.getType());
+    }
+
+    public void deleteTag(Long id) {
+        tagDAO.deleteById(id);
+    }
+
+    public TagDto findBySuggestion_Id(Long id) {
+       Tag tagOnSuggestion = tagDAO.findOneBySuggestion_Id(id);
+      return tagOnSuggestion==null?null: modelMapper.map(tagOnSuggestion, TagDto.class);}
+}
