--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/VotingSystemApplication.java
@@ -1,15 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@EnableDiscoveryClient
-@SpringBootApplication
-public class VotingSystemApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(VotingSystemApplication.class, args);
-    }
-
-}
