--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dao/MeetingDao.java
@@ -1,9 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dao;
-
-import cz.upce.votingsystemapplicationConfigServer.model.Meeting;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MeetingDao extends JpaRepository<Meeting, Long> {
-}
