--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dao/SuggestionDao.java
@@ -1,10 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dao;
-
-import cz.upce.votingsystemapplicationConfigServer.model.Suggestion;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SuggestionDao extends JpaRepository<Suggestion, Long> {
-
-}
