--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dao/UserDao.java
@@ -1,8 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dao;
-
-import cz.upce.votingsystemapplicationConfigServer.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserDao extends JpaRepository<User,Long> {
-
-}
