--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dao/VotingDao.java
@@ -1,12 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dao;
-
-import cz.upce.votingsystemapplicationConfigServer.model.Voting;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface VotingDao extends JpaRepository<Voting, Long> {
-    public List<Voting> getAllByUserId(Long userId);
-
-    public List<Voting> getAllBySuggestionId(Long suggestionId);
-}
