--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dto/SuggestionDto.java
@@ -1,24 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dto;
-
-public class SuggestionDto extends SuggestionForMeetingDto {
-
-  private MeetingForSuggestionDto meeting;
-
-  public SuggestionDto(){
-    super();
-  }
-
-  public SuggestionDto(long id, String content, MeetingForSuggestionDto meeting, Boolean accepted) {
-    super(id, content, accepted);
-
-    this.meeting = meeting;
-  }
-
-  public MeetingForSuggestionDto getMeeting() {
-    return meeting;
-  }
-
-  public void setMeeting(MeetingForSuggestionDto meeting) {
-    this.meeting = meeting;
-  }
-}
