--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dto/UserDto.java
@@ -1,49 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dto;
-
-public class UserDto {
-
-    private long id;
-    private String firstName;
-    private String lastName;
-    private String email;
-
-    public UserDto(long id, String firstName, String lastName, String email) {
-        this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.email = email;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-}
