--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/dto/VotingDto.java
@@ -1,33 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.dto;
-
-import cz.upce.votingsystemapplicationConfigServer.model.Voting;
-
-public class VotingDto {
-
-    private Voting.VOTE vote;
-    private SuggestionDto suggestion;
-//    TODO: Dodělat až bude
-//    private UserDto user;
-
-
-    public VotingDto(Voting.VOTE vote, SuggestionDto suggestion) {
-        this.vote = vote;
-        this.suggestion = suggestion;
-    }
-
-    public Voting.VOTE getVote() {
-        return vote;
-    }
-
-    public SuggestionDto getSuggestion() {
-        return suggestion;
-    }
-
-    public void setVote(Voting.VOTE vote) {
-        this.vote = vote;
-    }
-
-    public void setSuggestion(SuggestionDto suggestion) {
-        this.suggestion = suggestion;
-    }
-}
