--- voting-system-application/src/main/java/cz/upce/votingsystemapplicationConfigServer/service/TagService.java
@@ -1,42 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer.service;
-
-import cz.upce.votingsystemapplicationConfigServer.dao.TagDAO;
-import cz.upce.votingsystemapplicationConfigServer.dto.TagDto;
-import cz.upce.votingsystemapplicationConfigServer.model.Tag;
-import org.modelmapper.ModelMapper;
-import org.modelmapper.TypeToken;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class TagService {
-
-    @Autowired
-    private TagDAO tagDAO;
-
-    ModelMapper modelMapper=new ModelMapper();
-
-    public void addTag(Tag tag) {
-        tagDAO.save(tag);
-    }
-
-    public TagDto getTag(Long id) {
-        Tag tag = tagDAO.getOne(id);
-        return modelMapper.map(tag, TagDto.class);
-    }
-
-    public List<TagDto> getAllTag() {
-        List<Tag> all = tagDAO.findAll();
-        return modelMapper.map(all,new TypeToken<List<TagDto>>() {}.getType());
-    }
-
-    public void deleteTag(Long id) {
-        tagDAO.deleteById(id);
-    }
-
-    public TagDto findBySuggestion_Id(Long id) {
-       Tag tagOnSuggestion = tagDAO.findOneBySuggestion_Id(id);
-      return tagOnSuggestion==null?null: modelMapper.map(tagOnSuggestion, TagDto.class);}
-}
