+++ voting-system-application/src/test/java/cz/upce/votingsystemapplication/VotingSystemApplicationTests.java
@@ -1,0 +1,16 @@
+package cz.upce.votingsystemapplication;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class VotingSystemApplicationTests {
+
+    @Test
+    public void contextLoads() {
+    }
+
+}
