--- voting-system-application/src/test/java/cz/upce/votingsystemapplicationConfigServer/VotingSystemApplicationTests.java
@@ -1,16 +1,0 @@
-package cz.upce.votingsystemapplicationConfigServer;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class VotingSystemApplicationTests {
-
-    @Test
-    public void contextLoads() {
-    }
-
-}
